Move Label to io
