+++ app/geo-check/geo-check.cc
@@ -20,8 +20,8 @@
 #include "corecel/Assert.hh"
 #include "corecel/Types.hh"
 #include "corecel/cont/ArrayIO.json.hh"
-#include "corecel/cont/Label.hh"
 #include "corecel/cont/Range.hh"
+#include "corecel/io/Label.hh"
 #include "corecel/io/Logger.hh"
 #include "corecel/math/ArrayUtils.hh"
 #include "corecel/sys/Device.hh"
