+++ src/celeritas/geo/GeoParamsOutput.cc
@@ -19,7 +19,7 @@
 #if CELERITAS_USE_JSON
 #    include <nlohmann/json.hpp>
 
-#    include "corecel/cont/LabelIO.json.hh"
+#    include "corecel/io/LabelIO.json.hh"
 #    include "orange/BoundingBoxIO.json.hh"
 #endif
