+++ src/corecel/io/Label.cc
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file corecel/cont/Label.cc
+//! \file corecel/io/Label.cc
 //---------------------------------------------------------------------------//
 #include "Label.hh"
