+++ src/corecel/io/Label.hh
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file corecel/cont/Label.hh
+//! \file corecel/io/Label.hh
 //---------------------------------------------------------------------------//
 #pragma once
