+++ src/orange/OrangeParams.hh
@@ -12,10 +12,10 @@
 
 #include "corecel/OpaqueId.hh"
 #include "corecel/Types.hh"
-#include "corecel/cont/Label.hh"
 #include "corecel/cont/LabelIdMultiMap.hh"
 #include "corecel/cont/Span.hh"
 #include "corecel/data/CollectionMirror.hh"
+#include "corecel/io/Label.hh"
 
 #include "BoundingBox.hh"
 #include "OrangeData.hh"
