+++ src/orange/construct/OrangeInputIO.json.cc
@@ -17,10 +17,10 @@
 #include "corecel/Types.hh"
 #include "corecel/cont/Array.hh"
 #include "corecel/cont/ArrayIO.json.hh"
-#include "corecel/cont/Label.hh"
-#include "corecel/cont/LabelIO.json.hh"
 #include "corecel/cont/Range.hh"
 #include "corecel/cont/Span.hh"
+#include "corecel/io/Label.hh"
+#include "corecel/io/LabelIO.json.hh"
 #include "corecel/io/StringEnumMapper.hh"
 #include "orange/BoundingBoxIO.json.hh"
 #include "orange/OrangeTypes.hh"
